added report

+++ src/Main.java
@@ -1,10 +1,10 @@
 public class Main {
     public static void main(String[] args) {
-	int score = 2_000_000_000;
+	long score = 2_000_000_000;
 	System.out.println("score: " + score);
-	int receipt = 500_000_000;
+	long receipt = 500_000_000;
 	System.out.println("receipt: " + receipt);
-	int total = score + receipt;
+	long total = score + receipt;
 	System.out.println("total: " + total);
     }
 }
